+++ src/Symfony/Component/Translation/Tests/TranslatableTest.php
@@ -49,10 +49,8 @@ class TranslatableTest extends TestCase
 
     public function getTransTests()
     {
-        return [
-            ['Symfony est super !', new TranslatableMessage('Symfony is great!', [], ''), 'Symfony est super !', 'fr'],
-            ['Symfony est awesome !', new TranslatableMessage('Symfony is %what%!', ['%what%' => 'awesome'], ''), 'Symfony est %what% !', 'fr'],
-        ];
+        yield ['Symfony est super !', new TranslatableMessage('Symfony is great!', [], ''), 'Symfony est super !', 'fr'];
+        yield ['Symfony est awesome !', new TranslatableMessage('Symfony is %what%!', ['%what%' => 'awesome'], ''), 'Symfony est %what% !', 'fr'];
     }
 
     public function getFlattenedTransTests()